+++ apps/cold-api/src/platform/modules/resources/organizations/surveys/orgSurveys.service.ts
@@ -22,12 +22,12 @@ export class OrgSurveysService extends BaseWorker {
 
     if (latest?.status === survey_status_types.cold_submitted) {
       this.logger.error(`Survey ${surveyName} for ${orgId} already submitted`, { user, surveyName, orgId });
-      throw new UnprocessableEntityException(`Survey ${surveyName} for ${orgId} already submitted`);
+      throw new UnprocessableEntityException(`Unable to set status on ${surveyName}; already submitted by cold`);
     }
 
     if (latest?.status === survey_status_types.user_submitted && !user.isColdAdmin) {
       this.logger.error(`Survey ${surveyName} for ${orgId} already submitted by user`, { user, surveyName, orgId });
-      throw new UnprocessableEntityException(`Survey ${surveyName} for ${orgId} already submitted by user`);
+      throw new UnprocessableEntityException(`Unable to set status on ${surveyName}; already submitted by user`);
     }
 
     if (latest?.status === status) {
@@ -69,7 +69,7 @@ export class OrgSurveysService extends BaseWorker {
 
     const surveyId = surveyDef.id;
 
-    const surveyData = await this.prisma.survey_data.findFirst({
+    let surveyData = await this.prisma.survey_data.findFirst({
       where: {
         survey_definition_id: surveyDef.id,
         organization_id: orgId,
@@ -80,19 +80,28 @@ export class OrgSurveysService extends BaseWorker {
     });
 
     if (!surveyData) {
-      this.logger.error(`${surveyName} data for ${organization.name} not found`, {
+      this.logger.warn(`${surveyName} data for ${organization.name} not found`, {
         organization,
         user,
         survey: pick(surveyDef, ['id', 'name']),
       });
 
-      throw new NotFoundException(`${surveyName} data for ${organization.name} not found`);
+      surveyData = await this.prisma.survey_data.create({
+        data: {
+          id: new Cuid2Generator('sdata').scopedId,
+          survey_definition_id: surveyDef.id,
+          organization_id: orgId,
+          data: {},
+        },
+      });
+
+      //return [{ name: survey_status_types.draft, email: user.coldclimate_claims.email, created_at: new Date() }];
     }
 
     try {
       await this.prisma.survey_status.create({
         data: {
-          id: new Cuid2Generator('sustat').scopedId,
+          id: new Cuid2Generator('sstatus').scopedId,
           survey_id: surveyId,
           survey_name: surveyName,
           survey_data_id: surveyData.id,