+++ libs/react/src/components/pages/complianceWizard/complianceWizard.stories.tsx
@@ -1,6 +1,6 @@
 import { withKnobs } from '@storybook/addon-knobs';
 import { Meta, StoryObj } from '@storybook/react';
-import { getComplianceWizardFlowHandler, StoryMockProvider } from '@coldpbc/mocks';
+import { StoryMockProvider } from '@coldpbc/mocks';
 import { Route, Routes } from 'react-router-dom';
 import { ComplianceWizard, WizardRoutes } from '@coldpbc/components';
 
@@ -9,6 +9,13 @@ const meta: Meta<typeof ComplianceWizard> = {
   component: ComplianceWizard,
   tags: ['autodocs'],
   decorators: [withKnobs],
+  parameters: {
+    launchdarkly: {
+      flags: {
+        showNewCompliancePageHomeCold671: true,
+      },
+    },
+  },
 };
 
 export default meta;
@@ -17,7 +24,7 @@ type Story = StoryObj<typeof meta>;
 export const DocumentsUploadStep: Story = {
   render: args => {
     return (
-      <StoryMockProvider memoryRouterProps={{ initialEntries: ['/wizard/compliance/rei2/documents'] }}>
+      <StoryMockProvider memoryRouterProps={{ initialEntries: ['/wizard/compliance/rei_pia_2024/documents'] }}>
         <Routes>
           <Route path={'/compliance'} element={<div className={'text-tc-primary'}>Compliance Home</div>} />
           {WizardRoutes()}
@@ -31,21 +38,7 @@ export const DocumentsUploadStep: Story = {
 export const AutomateStep: Story = {
   render: args => {
     return (
-      <StoryMockProvider memoryRouterProps={{ initialEntries: ['/wizard/compliance/rei2/automate'] }}>
-        <Routes>
-          <Route path={'/compliance'} element={<div className={'text-tc-primary'}>Compliance Home</div>} />
-          {WizardRoutes()}
-          <Route path={'/assessments'} element={<div className={'text-tc-primary'}>Assessments</div>} />
-        </Routes>
-      </StoryMockProvider>
-    );
-  },
-};
-
-export const ProcessingStep: Story = {
-  render: args => {
-    return (
-      <StoryMockProvider memoryRouterProps={{ initialEntries: ['/wizard/compliance/rei2/processing'] }} handlers={getComplianceWizardFlowHandler.processing}>
+      <StoryMockProvider memoryRouterProps={{ initialEntries: ['/wizard/compliance/rei_pia_2024/automate'] }}>
         <Routes>
           <Route path={'/compliance'} element={<div className={'text-tc-primary'}>Compliance Home</div>} />
           {WizardRoutes()}
@@ -59,7 +52,7 @@ export const ProcessingStep: Story = {
 export const QuestionnaireStep: Story = {
   render: args => {
     return (
-      <StoryMockProvider memoryRouterProps={{ initialEntries: ['/wizard/compliance/rei2/questionnaire'] }}>
+      <StoryMockProvider memoryRouterProps={{ initialEntries: ['/wizard/compliance/rei_pia_2024/questionnaire'] }}>
         <Routes>
           <Route path={'/compliance'} element={<div className={'text-tc-primary'}>Compliance Home</div>} />
           {WizardRoutes()}